requirements.txt
click==6.7
colorlog==2.10.0
cycler==0.10.0
decorator==4.1.2
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
jupyter-client==5.1.0
jupyter-core==4.3.0
Keras==2.0.8
matplotlib==2.0.2
numpy==1.12.0
olefile==0.44
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
protobuf==3.4.0
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
requests==2.13.0
scipy==0.19.0
simplegeneric==0.8.1
six==1.10.0
tensorflow-gpu==1.0.0
tflearn==0.3
tornado==4.5.2
tqdm==4.11.2
traitlets==4.3.2
wcwidth==0.1.7
